+++ packages/amplify-graphql-index-transformer/jest.config.js
@@ -1,0 +1,12 @@
+const baseConfig = require('../../jest.config.base.js'); // eslint-disable-line @typescript-eslint/no-var-requires
+
+module.exports = {
+  ...baseConfig,
+  coverageThreshold: {
+    global: {
+      branches: 88,
+      lines: 90,
+      functions: 90,
+    },
+  },
+};