+++ packages/amplify-graphql-api-construct/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.18.1...@aws-amplify/graphql-api-construct@1.18.2) (2024-11-20)
+
+### Bug Fixes
+
+- bump ai-constructs to ^1.0.0 ([#3040](https://github.com/aws-amplify/amplify-category-api/issues/3040)) ([0e9e195](https://github.com/aws-amplify/amplify-category-api/commit/0e9e195b05e7d40e41c29b82b8f3fb3a8961d80b))
+
+## [1.18.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.18.0...@aws-amplify/graphql-api-construct@1.18.1) (2024-11-20)
+
+**Note:** Version bump only for package @aws-amplify/graphql-api-construct
+
 # [1.18.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.17.3...@aws-amplify/graphql-api-construct@1.18.0) (2024-11-19)
 
 ### Features